Separate host and storage in separate crates as well
