+++ crates/sdk-host-mock/Cargo.toml
@@ -1,5 +1,5 @@
 [package]
-name = "svm-sdk-mock"
+name = "svm-sdk-host-mock"
 version = "0.0.0"
 authors = ["Spacemesh SVM Team"]
 license = "MIT"
@@ -11,38 +11,29 @@ description = "Spacemesh Virtual Machine"
 publish = false
 
 [dependencies]
-lazy_static = "1.4.0"
 svm-sdk-alloc = { path = "../sdk-alloc", default-features = false }
+svm-sdk-host = { path = "../sdk-host" }
 svm-sdk-types = { path = "../sdk-types", default-features = false }
 svm-sdk-std = { path = "../sdk-std", default-features = false }
-svm-sdk-host = { path = "../sdk-host" }
-svm-sdk-host-mock = { path = "../sdk-host-mock" }
-svm-sdk-storage = { path = "../sdk-storage" }
-svm-sdk-storage-mock = { path = "../sdk-storage-mock" }
-svm-sdk-macros = { path = "../sdk-macros", default-features = false }
 svm-abi-encoder = { path = "../abi-encoder", default-features = false }
 svm-abi-decoder = { path = "../abi-decoder", default-features = false }
 
+[dev-dependencies]
+lazy_static = "1.4.0"
+
 [features]
-default = ["meta"]
-meta = ["svm-sdk-macros/meta"]
+mock = []
 static-alloc = [
-    "svm-sdk-alloc/static-alloc",
-    "svm-sdk-types/static-alloc",
+    "svm-sdk-alloc/static-alloc", 
+    "svm-sdk-types/static-alloc", 
     "svm-sdk-std/static-alloc",
-    "svm-sdk-host-mock/static-alloc",
-    "svm-sdk-storage-mock/static-alloc",
-    "svm-sdk-macros/static-alloc",
     "svm-abi-encoder/static-alloc",
     "svm-abi-decoder/static-alloc",
 ]
 dynamic-alloc = [
-    "svm-sdk-alloc/dynamic-alloc",
-    "svm-sdk-types/dynamic-alloc",
+    "svm-sdk-alloc/dynamic-alloc", 
+    "svm-sdk-types/dynamic-alloc", 
     "svm-sdk-std/dynamic-alloc",
-    "svm-sdk-host-mock/dynamic-alloc",
-    "svm-sdk-storage-mock/dynamic-alloc",
-    "svm-sdk-macros/dynamic-alloc",
     "svm-abi-encoder/dynamic-alloc",
     "svm-abi-decoder/dynamic-alloc",
 ]
