+++ crates/sdk-storage-mock/Cargo.toml
@@ -13,6 +13,7 @@ publish = false
 [dependencies]
 lazy_static = "1.4.0"
 svm-sdk-alloc = { path = "../sdk-alloc", default-features = false }
+svm-sdk-storage = { path = "../sdk-storage" }
 svm-sdk-types = { path = "../sdk-types", default-features = false }
 svm-sdk-std = { path = "../sdk-std", default-features = false }
