+++ crates/sdk/Cargo.toml
@@ -15,11 +15,16 @@ lazy_static = "1.4.0"
 svm-sdk-alloc = { path = "../sdk-alloc", default-features = false }
 svm-sdk-types = { path = "../sdk-types", default-features = false }
 svm-sdk-std = { path = "../sdk-std", default-features = false }
-svm-sdk-host = { path = "../sdk-host", features = ["ffi" ]}
-svm-sdk-storage = { path = "../sdk-storage", features = ["ffi"] }
+svm-sdk-host = { path = "../sdk-host" }
+svm-sdk-host-ffi = { path = "../sdk-host-ffi" }
+svm-sdk-storage = { path = "../sdk-storage" }
+svm-sdk-storage-ffi = { path = "../sdk-storage-ffi" }
 svm-sdk-macros = { path = "../sdk-macros", default-features = false }
 svm-abi-encoder = { path = "../abi-encoder", default-features = false }
 svm-abi-decoder = { path = "../abi-decoder", default-features = false }
+
+[dev-dependencies]
+svm-sdk = { path = ".", features = ["dynamic-alloc"] }
 
 [features]
 default = ["meta"]
@@ -28,8 +33,6 @@ static-alloc = [
     "svm-sdk-alloc/static-alloc",
     "svm-sdk-types/static-alloc",
     "svm-sdk-std/static-alloc",
-    "svm-sdk-host/static-alloc",
-    "svm-sdk-storage/static-alloc",
     "svm-sdk-macros/static-alloc",
     "svm-abi-encoder/static-alloc",
     "svm-abi-decoder/static-alloc",
@@ -38,8 +41,6 @@ dynamic-alloc = [
     "svm-sdk-alloc/dynamic-alloc",
     "svm-sdk-types/dynamic-alloc",
     "svm-sdk-std/dynamic-alloc",
-    "svm-sdk-host/dynamic-alloc",
-    "svm-sdk-storage/dynamic-alloc",
     "svm-sdk-macros/dynamic-alloc",
     "svm-abi-encoder/dynamic-alloc",
     "svm-abi-decoder/dynamic-alloc",
